WIP 11: add STDERR log entry handling
