+++ Gopkg.lock
@@ -11,7 +11,8 @@
   name = "github.com/apex/log"
   packages = [
     ".",
-    "handlers/cli"
+    "handlers/cli",
+    "handlers/memory"
   ]
   revision = "0296d6eb16bb28f8a0c55668affcf4876dc269be"
   version = "v1.0.0"
@@ -368,6 +369,6 @@
 [solve-meta]
   analyzer-name = "dep"
   analyzer-version = 1
-  inputs-digest = "cef0cf926e66becbcaeec0ac04638120df41cf807fe1afa82fafc3b28d895d6d"
+  inputs-digest = "80c8e5176465abf45825ca7a607383ff2f2f95fa425af7dce3686a575c048627"
   solver-name = "gps-cdcl"
   solver-version = 1
