+++ cmd/fossa/display/progress.go
@@ -1,6 +1,7 @@
 package display
 
 import (
+	"github.com/apex/log"
 	"github.com/briandowns/spinner"
 )
 
@@ -11,7 +12,7 @@ var (
 
 // InProgress shows a progress spinner with a message.
 func InProgress(message string) {
-	if useSpinner {
+	if useSpinner && level > log.DebugLevel {
 		s.Suffix = " " + message
 		s.Restart()
 	}
